requirements.txt
numpy
matplotlib
pandas
scikit-learn
torch
transformers
openpyxl
tqdm
ollama
python-dotenv
